notes are now retreived from database based on the topic_id
api has been updated to create topics and notes. Soon delete notes and topics with the help of a dynamic Delete button.

information is being retreived and displayed in a Base component that only contains the main info (name, title, description , etc) a Alltopics, notes files, styles and formats the cards using tailwindcss then that is used in the main page.